Moved the organizer to models.py to avoid circular import. Opening the bug report now
