+++ topchef/models.py
@@ -4,9 +4,12 @@ have JSON representations written in marshmallow, and a single representation
 in the database.
 """
 import os
+import tempfile
 import uuid
 import logging
 import json
+from uuid import UUID
+
 import jsonschema
 from datetime import datetime, timedelta
 from flask import url_for
@@ -17,11 +20,75 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import Session, relationship
 from . import database
 from .config import config
-from .schema_directory_organizer import SchemaDirectoryOrganizer
 
 LOG = logging.getLogger(__name__)
 
 BASE = declarative_base(metadata=database.METADATA)
+
+
+class SchemaDirectoryOrganizer(object):
+    REGISTRATION_SCHEMA_NAME = 'job_registration_schema'
+    RESULT_SCHEMA_NAME = 'job_result_schema'
+
+    def __init__(self, schema_directory_path):
+        self.root_path = schema_directory_path
+
+    @property
+    def services(self):
+        return [
+            service_id for service_id in os.listdir(self.root_path)
+                if self._is_guid(service_id)
+            ]
+
+    def register_service(self, service):
+        service_directory = os.path.join(self.root_path, str(service.id))
+
+        os.mkdir(service_directory)
+
+    def __getitem__(self, model):
+        if isinstance(model, Service):
+            return os.path.join(
+                self.root_path, str(model.id)
+            )
+        elif isinstance(model, Job):
+            return os.path.join(
+                self.root_path, str(model.parent_service.id),
+                '%s.json' % str(model.id)
+            )
+        else:
+            raise ValueError(
+                'The model class %s is not a Service or Job',
+                model.__repr__()
+            )
+
+    def write(self, data_to_write, target_path):
+
+        _, temporary_filename = tempfile.mkstemp(suffix='.json')
+
+        with open(temporary_filename, mode='w') as temporary_file:
+            temporary_file.write(data_to_write)
+
+        if os.path.isfile(target_path):
+            os.remove(target_path)
+
+        os.rename(temporary_file.name, target_path)
+
+        if os.path.isfile(temporary_filename):
+            os.remove(temporary_filename)
+
+    @staticmethod
+    def _is_guid(dirname):
+        try:
+            UUID(dirname)
+            return True
+        except ValueError:
+            return False
+
+    def __repr__(self):
+        return '%s(schema_directory_path=%s)' % (
+            self.__class__.__name__, self.root_path
+        )
+
 
 FILE_MANAGER = SchemaDirectoryOrganizer(config.SCHEMA_DIRECTORY)
